+++ .gitignore
@@ -1,3 +1,6 @@
 .idea/*
 venv/
 .vscode
+lib/__pycache__/
+.DS_Store
+lib/.DS_Store
